requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
colour==0.1.5
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Deprecated==1.2.13
dnspython==2.2.0
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.3.0
gql==3.0.0
graphql-core==3.2.0
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lxml==4.7.1
MarkupSafe==2.0.1
multidict==6.0.2
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
PyGithub==1.55
PyJWT==2.3.0
pymongo==4.0.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
Pyrogram==1.3.6
PySocks==1.7.1
python-dateutil==2.8.2
queuelib==1.6.2
regex==2022.1.18
requests==2.27.1
requests-toolbelt==0.9.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
TgCrypto==1.2.3
toml==0.10.2
Twisted==21.7.0
typing-extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
websockets==10.1
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zope.interface==5.4.0
